# swag_docs/swagger.yaml
definitions:
  AdminAccountJoined:
    properties:
      created_at:
        type: string
      created_by_name:
        type: string
      created_by_urn:
        type: string
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      email_verified_at_ts:
        type: integer
      external_id:
        type: string
      first_name:
        type: string
      hashed_password:
        type: string
      id:
        type: string
      is_super_user_session:
        type: integer
      last_login_ts:
        type: integer
      last_name:
        type: string
      name:
        type: string
      organization_id:
        type: string
      organization_name:
        type: string
      password_updated_at_ts:
        type: integer
      phone:
        type: string
      properties:
        $ref: '#/definitions/account.Properties'
      role:
        type: integer
      signup_properties:
        $ref: '#/definitions/account.SignupProperties'
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        type: string
      updated_by_name:
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
    type: object
...
  account.AccountJoinedPublic:
    properties:
      created_at:
        format: date-time
        type: string
      created_by_urn:
        type: string
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      external_id:
        type: string
      first_name:
        type: string
      id:
        format: uuid
        type: string
      is_super_user_session:
        type: integer
      last_name:
        type: string
      organization_id:
        format: uuid
        type: string
      phone:
        type: string
      role:
        type: integer
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        format: date-time
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
    required:
    - id
    - urn
    - created_by_urn
    - updated_by_urn
    - status
    - deleted
    - disabled
    - updated_at
    - created_at
    - first_name
    - last_name
    - email
    - phone
    - external_id
    - test_user_type
    - organization_id
    - role
    - is_super_user_session
    type: object
  account.AccountPublic:
    properties:
      created_at:
        format: date-time
        type: string
      created_by_urn:
        type: string
      deleted:
        type: integer
      disabled:
        type: integer
      email:
        type: string
      external_id:
        type: string
      first_name:
        type: string
      id:
        format: uuid
        type: string
      is_super_user_session:
        type: integer
      last_name:
        type: string
      organization_id:
        format: uuid
        type: string
      phone:
        type: string
      role:
        type: integer
      status:
        type: integer
      test_user_type:
        type: integer
      updated_at:
        format: date-time
        type: string
      updated_by_urn:
        type: string
      urn:
        type: string
    required:
    - id
    - urn
    - created_by_urn
    - updated_by_urn
    - status
    - deleted
    - disabled
    - updated_at
    - created_at
    - first_name
    - last_name
    - email
    - phone
    - external_id
    - test_user_type
    - organization_id
    - role
    - is_super_user_session
    type: object
...
  account.Properties:
    properties:
      external_user_info: {}
      invite_key:
        type: string
      invite_ts:
        format: int64
        type: integer
      last_seen:
        format: int64
        type: integer
      verify_email_key:
        type: string
    type: object
...
  account.SignupProperties:
    properties:
      is_oauth:
        format: int64
        type: integer
    required:
    - is_oauth
    type: object
...
  accounts.CheckKeyResponse:
    properties:
      valid:
        type: boolean
    type: object
...
  accounts.PasswordResetInput:
    properties:
      email:
        type: string
      hash:
        type: string
    type: object
...
  accounts.ResetPasswordInput:
    properties:
      password:
        type: string
      password_confirmation:
        type: string
      resetKey:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /account/{id}:
    delete:
      consumes:
      - application/json
      description: Marks an account as deleted by setting status to USER_DELETED
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/AdminAccountJoined'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete account
      tags:
      - Account
    get:
      consumes:
      - application/json
      description: Get Account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.AccountJoinedPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get Account
      tags:
      - Account
    put:
      consumes:
      - application/json
      description: Update Account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/account.AccountPublic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.AccountPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update Account
      tags:
      - Account
  /account/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancels a pending invitation and invalidates the invite session
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/AdminAccountJoined'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Cancel invite
      tags:
      - Account
  /account/{id}/resend:
    post:
      consumes:
      - application/json
      description: Resends invitation email to a pending account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/account.AccountJoinedPublic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Resend invite
      tags:
      - Account
...
  /password/check:
    get:
      consumes:
      - application/json
      description: Validates whether a password reset key is still valid and active
      parameters:
      - description: Reset key to validate
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/accounts.CheckKeyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Check reset key validity
      tags:
      - Account
  /password/reset:
    post:
      consumes:
      - application/json
      description: Resets the user's password using the reset key from email
      parameters:
      - description: Password reset details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.ResetPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Reset password
      tags:
      - accounts
  /password/reset/send:
    post:
      consumes:
      - application/json
      description: Sends a password reset email with a temporary session key
      parameters:
      - description: Email and verification hash
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/accounts.PasswordResetInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Send password reset email
      tags:
      - accounts
...
swagger: "2.0"